Emplace entities into cells
The map is subdivided in cells, where each cell has a specific position
for a tile. The emplace method should be used to put entities on the
map. A collecion of entities is maintained by the map to ensure an old
entity is replaced by a new one if it is on the same cell. Then, add a
player action to put some tiles on the map on mouse click.